Version2/pyproject.toml
[project]
name = "ai-data-analyst"
version = "0.1.0"
description = "AI Data Analyst with natural language to SQL"
requires-python = "==3.12"
dependencies = [
    "langgraph>=0.0.55",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.1.0",
    "openai>=1.10.0",
    "sqlalchemy>=2.0.23",
    "pandas>=2.1.3",
    "plotly>=5.17.0",
    "streamlit>=1.28.1",
    "psycopg>=3.1.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.6.0",
    "pymupdf>=1.23.0",
    "tiktoken>=0.5.0",
]

[tool.uv]
package = true

[project.scripts]
start = "streamlit:main"

[tool.setuptools]
py-modules = ["config", "database", "database_test", "streamlit_app", "visualization", "agent", "llm_handler", "models"]
